--- .gitignore
@@ -1,2 +1,0 @@
-# Ignore all .exe files
-*.exe
