+++ Cell.hpp
@@ -1,13 +1,22 @@
 #pragma once
 #include "Mark.hpp"
 
+
 class Cell
 {
 private:
     Mark state;
 
 public:
-    Cell();
-    void setMark(Mark L);
-    Mark getMark() const;
+    Cell() : state(Mark::Gol) {}
+
+    void setMark(Mark L)
+    {
+        state = L;
+    }
+
+    Mark getMark() const
+    {
+        return state;
+    }
 };
